Rewrite everything
The previous iteration of this module wanted templates for each section
of the pf.conf file.  This quickly became unusable and less flexible.

This rewrite instead leaves the power in the users hands for how she
wishes to use PF, loosing no flexibility, and instead gaining the
power of managing rules with ERB templates, combining the power of
Puppet, ERB and PF for a knockout combo... at least thats the theory.