Add pf::table define for class selection
The list of classes passed is used as the argument in a PuppetDB query
to get a list of IPaddresses for nodes that contain the requeste class.
Those IPs are added to the table dynamically.  This feature means tables
can be updated automatically based on node classification and can be
used in the main pf class template.

Future support here will likely include other items to query, and some
restrictions on which address family is being searched.