Improve flexibility of table population
Here we add logic to allow for querying a common class, and using a parameter
value of the returned class as the value with which to populate a table address.
This allows table population when the IP desired is not the value of the
ipaddress or ipaddress6 fact, allowing each node to override in hiera with the
value on a common class.